refactor: create specific request interfaces
