+++ src/authentication/authentication.controller.ts
@@ -1,7 +1,7 @@
 import { Controller, Post, Req, UseGuards } from '@nestjs/common';
-import { AuthenticationService } from './authentication.service';
-import { Request } from 'express';
 import { LocalAuthGuard } from 'src/common/guards/local-auth.guard';
+import { AuthenticationService } from './authentication.service';
+import { UserModelRequest } from 'src/common/interfaces';
 
 @Controller('api/auth')
 export class AuthenticationController {
@@ -9,7 +9,7 @@ export class AuthenticationController {
 
   @UseGuards(LocalAuthGuard)
   @Post('login')
-  public async login(@Req() req: Request) {
+  public async login(@Req() req: UserModelRequest) {
     return this.service.login(req.user);
   }
 }
