--- src/common/@types/Express.d.ts
@@ -1,8 +1,0 @@
-import type { Request } from 'express';
-import { UserModel } from 'src/models/user.model';
-
-declare module 'express' {
-  export interface Request {
-    user: UserModel;
-  }
-}
