+++ src/common/interfaces/index.ts
@@ -1,0 +1,3 @@
+export * from './payload-jwt-request.interface'
+export * from './user-model-request.interface'
+export * from './payload-jwt.interface'
