+++ src/common/interfaces/payload-jwt-request.interface.ts
@@ -1,0 +1,6 @@
+import { Request } from 'express';
+import { PayloadJWT } from './payload-jwt.interface';
+
+export interface PayloadJWTRequest extends Request {
+  user: PayloadJWT;
+}
