+++ src/common/interfaces/user-model-request.interface.ts
@@ -1,0 +1,6 @@
+import { Request } from 'express';
+import { UserModel } from 'src/models/user.model';
+
+export interface UserModelRequest extends Request {
+  user: UserModel;
+}
